feat(api): SSE ingest progress, file preview, chat history, web-rag overhaul
* runners: add ingestion_runner (SSE progress for file ingest), add
  risk_predict_runner, drop regulation_runner (folded into web_rag).
  Existing runners pick up the new EventBus envelope shape.
* services: add history (multi-turn chat history retrieval) and
  preview (lazy pypdfium2 page rendering); files service grows the
  upload + purge + preview facade; graph_service gains sample queries
  and admin filters; web_rag is rewritten around the Tavily client.
* routes: admin trace detail / users; chat with web mode + history;
  files preview + ingest progress streaming; insurance workbench
  routes adjust to the new schemas.
* main: lifespan-cached heavy resources, alembic-aware startup.
* README: surface the new web layer at a glance.

+++ src/api/schemas/admin.py
@@ -13,12 +13,12 @@ class ConfigEntrySchema(BaseModel):
     """One row in the admin UI's config table."""
 
     key: str
-    type: str                # "int" | "str"
+    type: str                # "int" | "str" | "float" | "bool"
     default: Any
     description: str = ""
     group: str = ""
-    min: Optional[int] = None
-    max: Optional[int] = None
+    min: Optional[float] = None
+    max: Optional[float] = None
     min_length: Optional[int] = None
     max_length: Optional[int] = None
